Add files via upload
Added functionality so that multiple modules could be created. Also made it so that if a module calls a module, it saves its instance as a part of that module. Next steps will be to finish part2gate method so that a module may be represented at the gate level(ORs, ANDs, XORs)