Add files via upload
part2gate implementation for comparator and multiplier complete. It has been decided that a divider will probably not be implemented for this program as it is very complicated, impractical in most cases, and would be very difficult to dynamically create. There is potential for hard-coded versions of the divider up to 4 bits, but this would be the extent to what we are capable of completing given the scope of this project. TODO: part2gate for Mux, setting up for single and multiple & logic as well as | logic